LayoutTests/webaudio/oscillatornode-legacy-api.html

<!DOCTYPE html>
<html>
<head>
    <script src="../resources/js-test.js"></script>
</head>
<body>
    <script>
        description("Testing that the legacy OscillatorNode API is not exposed on the unprefixed OscillatorNode");

        shouldBeUndefined("OscillatorNode.prototype.playbackState");

        shouldBeUndefined("OscillatorNode.UNSCHEDULED_STATE");
        shouldBeUndefined("OscillatorNode.SCHEDULED_STATE");
        shouldBeUndefined("OscillatorNode.PLAYING_STATE");
        shouldBeUndefined("OscillatorNode.FINISHED_STATE");
        shouldBeUndefined("OscillatorNode.prototype.UNSCHEDULED_STATE");
        shouldBeUndefined("OscillatorNode.prototype.SCHEDULED_STATE");
        shouldBeUndefined("OscillatorNode.prototype.PLAYING_STATE");
        shouldBeUndefined("OscillatorNode.prototype.FINISHED_STATE");

        let context = new AudioContext;
        let oscillator = context.createOscillator();
        shouldBeUndefined("oscillator.playbackState");
        shouldBeUndefined("oscillator.UNSCHEDULED_STATE");
        shouldBeUndefined("oscillator.SCHEDULED_STATE");
        shouldBeUndefined("oscillator.PLAYING_STATE");
        shouldBeUndefined("oscillator.FINISHED_STATE");
        shouldBeUndefined("oscillator.__proto__.playbackState");
        shouldBeUndefined("oscillator.__proto__.UNSCHEDULED_STATE");
        shouldBeUndefined("oscillator.__proto__.SCHEDULED_STATE");
        shouldBeUndefined("oscillator.__proto__.PLAYING_STATE");
        shouldBeUndefined("oscillator.__proto__.FINISHED_STATE");
    </script>
</body>
</html>
